// test/utils/enums.ts
export enum OS {
	ANDROID = "android",
	IOS = "ios",
	WEB = "web"
}

export enum Brand {
	KK = "kayak",
	MM = "momondo",
	HC = "hotelscombined",
	SW = "swoodoo",
	CX = "checkfelix",
	CF = "cheapflights",
	MU = "mundi",
}

export enum Locale {
	en_US,
	da_DK,
	ko_KR,
	de_DE,
	fr_FR,
	de_AT,
	es_ES,
	en_GB,
	zh_CN,
	it_IT,
	sv_SE,
}

export enum By {
	Predicate = "-ios predicate string:",
	Accessiblity = "~",
	ClassChain = "-ios class chain:",
	ID = "id=",
	UiAutomator = "android="
}